+++ flutter/lib/src/native/cocoa/binding.dart
@@ -22996,6 +22996,8 @@ class SentryCocoa {
       _registerName1("prefersIncrementalDelivery");
   late final _sel_setPrefersIncrementalDelivery_1 =
       _registerName1("setPrefersIncrementalDelivery:");
+  late final _sel_getGraphQLOperationName1 =
+      _registerName1("getGraphQLOperationName");
   late final _sel_storeCookies_forTask_1 =
       _registerName1("storeCookies:forTask:");
   void _objc_msgSend_804(
@@ -37671,7 +37673,8 @@ class ObjCBlock_bool_ObjCObject_ffiUnsignedLong_bool extends _ObjCBlockBase {
   ObjCBlock_bool_ObjCObject_ffiUnsignedLong_bool.fromFunctionPointer(
       SentryCocoa lib,
       ffi.Pointer<
-              ffi.NativeFunction<
+              ffi
+              .NativeFunction<
                   ffi.Bool Function(ffi.Pointer<ObjCObject> arg0,
                       ffi.UnsignedLong arg1, ffi.Pointer<ffi.Bool> arg2)>>
           ptr)
@@ -42115,15 +42118,17 @@ class ObjCBlock_bool_ObjCObject_bool extends _ObjCBlockBase {
                       ffi.Pointer<ffi.Bool> arg1)>>
           ptr)
       : this._(
-            lib._newBlock1(
-                _cFuncTrampoline ??= ffi.Pointer.fromFunction<
-                            ffi.Bool Function(
-                                ffi.Pointer<_ObjCBlock> block,
-                                ffi.Pointer<ObjCObject> arg0,
-                                ffi.Pointer<ffi.Bool> arg1)>(
-                        _ObjCBlock_bool_ObjCObject_bool_fnPtrTrampoline, false)
-                    .cast(),
-                ptr.cast()),
+            lib
+                ._newBlock1(
+                    _cFuncTrampoline ??= ffi.Pointer.fromFunction<
+                                ffi.Bool Function(
+                                    ffi.Pointer<_ObjCBlock> block,
+                                    ffi.Pointer<ObjCObject> arg0,
+                                    ffi.Pointer<ffi.Bool> arg1)>(
+                            _ObjCBlock_bool_ObjCObject_bool_fnPtrTrampoline,
+                            false)
+                        .cast(),
+                    ptr.cast()),
             lib);
   static ffi.Pointer<ffi.Void>? _cFuncTrampoline;
 
@@ -58356,6 +58361,11 @@ class NSURLSessionTask extends NSObject {
     final _ret =
         _lib._objc_msgSend_2(_lib._class_NSURLSessionTask1, _lib._sel_new1);
     return NSURLSessionTask._(_ret, _lib, retain: false, release: true);
+  }
+
+  NSString getGraphQLOperationName() {
+    final _ret = _lib._objc_msgSend_20(_id, _lib._sel_getGraphQLOperationName1);
+    return NSString._(_ret, _lib, retain: true, release: true);
   }
 
   static NSURLSessionTask allocWithZone_(